feat(ux): split tabbed /tasks hub into dedicated pages
The legacy /tasks page hosted four tabs (Repo Tasks / Standalone / Issues /
PRs) competing with the sidebar. Now that the sidebar exposes Jobs, Reviews,
and Issues as first-class destinations, the tab layer is redundant.

- /tasks now renders just the Repo Task list (no tabs, focused header)
- /jobs becomes a real list page using StandaloneList instead of redirecting
- /reviews becomes a real list page using PrBrowser instead of redirecting
- /issues is new, hosting the IssuesBrowser extracted from /tasks
- Sidebar gains Issues under Run; final shape is
    Run    Tasks · Jobs · Reviews · Issues · Scheduled
    Live   Agents · Sessions
- Back-compat: /tasks?tab=standalone|issues|prs client-side redirects to the
  new dedicated pages so existing bookmarks keep working.
- Each new page has a one-line subtitle telling the user what the section is
  for, since the sidebar label alone isn't always self-explanatory.

The IssuesBrowser was inlined as a 220-line function inside /tasks/page.tsx;
extracted to apps/web/src/components/issues-browser.tsx as its own component.
